feat: Export milvusclient.annRequest (issue: #41261) (#41356)
Issue: #41261

`milvusclient.NewHybridSearchOption` receives a variadic `annRequests`
parameter. However, since `milvusclient.annRequest` is private, there is
no way to declare a slice, therefore there is no way to make it fully
generic (as in, create a slice of `milvusclient.annRequest`s and pass
them to `NewHybridSearchOption`. This PR renames
`milvusclient.annRequest` to `milvusclient.AnnRequest` to export it.

This is an API change since it's renaming a struct. However, since the
struct was previously private no external code depends on it, unless
it's doing nasty things with reflection (in which case it should not
depend on the name).

Signed-off-by: Mario Camou <mcamou@users.noreply.github.com>

Signed-off-by: mcamou <mcamou@users.noreply.github.com>